관측지 TODO 수정
