+++ app/src/main/java/com/example/byeoldori/ui/components/observatory/MarkerCard.kt
@@ -1,32 +1,18 @@
 package com.example.byeoldori.ui.components.observatory
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Place
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
+import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
+import androidx.compose.ui.*
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.*
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.byeoldori.ui.theme.Blue500
-import com.example.byeoldori.ui.theme.Blue800
-import com.example.byeoldori.ui.theme.Purple500
-import com.example.byeoldori.ui.theme.Purple800
-import com.example.byeoldori.ui.theme.TextHighlight
+import androidx.compose.ui.unit.*
+import com.example.byeoldori.ui.theme.*
 
 @Composable
 fun MarkerCardWithGradient() {
