+++ app/src/main/java/com/example/byeoldori/viewmodel/Observatory/CurrentWeather.kt
@@ -1,0 +1,8 @@
+package com.example.byeoldori.viewmodel.Observatory
+
+data class CurrentWeather(
+    val temperature: String,
+    val humidity: String,
+    val windSpeed: String,
+    val suitability: String
+)
