+++ app/src/main/java/com/example/byeoldori/viewmodel/Observatory/ObservatoryModel.kt
@@ -1,0 +1,19 @@
+package com.example.byeoldori.viewmodel.Observatory
+
+import com.naver.maps.geometry.LatLng
+
+enum class ObservatoryType {
+    GENERAL, POPULAR
+}
+
+data class Observatory(
+    val name: String,
+    val type: ObservatoryType,
+    val latLng: LatLng,
+    val reviewCount: Int,
+    val likeCount: Int,
+    val avgRating: Float,
+    val address: String,
+    val imageRes: Int,
+    val suitability: Int // 예: 87%
+)
